Proposed fix for reap-cycle cycling
Added in a cache for our wrapped Puma::Worker objects so that whenever
we call PumaAutoTune::Master#workers we will always get the same
PumaAutoTune::Worker objects. This lets the concept of restarting a
worker persist through calls to reset on PumaAutoTune::Memory. We want
resetting the memory to recalculate the memory EXCEPT for the fact that
we want the restarting worker to still return 0.